requirements/base.txt
pytz==2023.3  # https://github.com/stub42/pytz
python-slugify==8.0.1  # https://github.com/un33k/python-slugify
Pillow==9.5.0  # https://github.com/python-pillow/Pillow
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.4.0  # https://github.com/evansd/whitenoise
redis==4.5.5  # https://github.com/redis/redis-py
hiredis==2.2.3  # https://github.com/redis/hiredis-py
pyotp==2.8.0  # https://github.com/pyauth/pyotp
cerberus==1.3.4  # https://github.com/pyeve/cerberus
phonenumbers  # https://github.com/daviddrysdale/python-phonenumbers
# Django
# ------------------------------------------------------------------------------
django==4.1.9  # pyup: < 4.2  # https://www.djangoproject.com/
django-environ==0.10.0  # https://github.com/joke2k/django-environ
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.54.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.7  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.2.0  # https://github.com/jazzband/django-redis
django-simple-history==3.3.0 # https://github.com/jazzband/django-simple-history
django-turnstile==0.1.0  # https://gist.github.com/mikeckennedy/66d6298106671a3c6215c9262c102d74
django-phonenumber-field==7.1.0  # https://github.com/stefanfoulis/django-phonenumber-field
django-storages[boto3]==1.13.2  # https://github.com/jschneier/django-storages
django-ckeditor==6.5.1 # https://github.com/django-ckeditor/django-ckeditor
django-widget-tweaks==1.4.12

# Other
stripe==5.4.0

# Database
# This is the new for 4.2 you need both the binary and package.
psycopg-binary==3.1.9
psycopg==3.1.9

# old db for 4.1.8
psycopg2-binary==2.9.6
